Moved posts code from User into its own component
