+++ client/src/App.css
@@ -43,7 +43,7 @@ form {
 }
 
 
-form input {
+form input, form textarea {
     font-family: 'Acme', sans-serif;
     background: #666666;
     border: none;
@@ -51,6 +51,12 @@ form input {
     padding: 0.5rem;
     outline: none;
     border-radius: 3px;
+}
+
+textarea {
+    width: 100%;
+    height: 12rem;
+    font-size: 2rem;
 }
 
 ::placeholder {
